pyproject.toml
[tool.poetry]
name = "subscription service"
version = "0.1.0"
description = ""
authors = ["Vlad <ruskiborg@list.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.1"
gunicorn = "^23.0.0"
uvicorn = "^0.30.6"
python-dotenv = "^1.0.1"
psycopg = {extras = ["binary"], version = "^3.2.2"}
celery = {extras = ["redis"], version = "^5.4.0"}
django-celery-beat = "^2.7.0"
django-extensions = "^3.2.3"
djangorestframework = "^3.15.2"
drf-spectacular = "^0.27.2"
drf-spectacular-sidecar = "^2024.7.1"
djangorestframework-dataclasses = "^1.3.1"
django-filter = "^23.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
